Pull request #3626: [BLUETOOTH-15739] Applications: Fix Cert C issues
Squashed commit of the following:

commit a55613bbc4482489f7b6ac3d7cf5d0eb60d5ffe0
Author: Cosmin Grumei <cosmin-teodor.grumei@nxp.com>
Date:   Thu Aug 21 17:25:03 2025 +0300

    [BLUETOOTH-15739] Fix for bus fault caused by uint16_t value written as uint8_t and read as uint16_t.

commit bb63d076687f1f6925761736a005e7ff0837eea6
Author: Cosmin Grumei <cosmin-teodor.grumei@nxp.com>
Date:   Thu Aug 21 12:13:35 2025 +0300

    [BLUETOOTH-15739] Revert NULL pointer dereferencing fix

commit e587f6ac33fcf78c39cc6da89af816abda5fedd1
Author: Cosmin Grumei <cosmin-teodor.grumei@nxp.com>
Date:   Wed Aug 20 17:18:14 2025 +0300

    [BLUETOOTH-15739] Use NumberOfElements macro

commit 0091fc3ed4dce82591113e7229fcce5926bfc5ef
Author: Cosmin Grumei <cosmin-teodor.grumei@nxp.com>
Date:   Wed Aug 20 15:32:48 2025 +0300

    [BLUETOOTH-15739] Redone NULL pointer dereferencing fix to avoid MISRA issue

commit f50e664c8ccf9458de8dd42516ff907a862cc866
Author: Cosmin Grumei <cosmin-teodor.grumei@nxp.com>
Date:   Tue Aug 19 18:08:41 2025 +0300

    [BLUETOOTH-15739] Applications: Fix Cert C issues

commit 1c12bef90c87b14dfa76a52ae25ea34bd4a93e56
Author: Cosmin Grumei <cosmin-teodor.grumei@nxp.com>
Date:   Wed Aug 13 19:23:10 2025 +0300

    [BLUETOOTH-15739] Applications: Fix Cert C issues

commit 149923f93e45078d622d9f97bbc17f4fb3c73f58
Author: Cosmin Grumei <cosmin-teodor.grumei@nxp.com>
Date:   Mon Aug 11 17:48:21 2025 +0300

    [BLUETOOTH-15739] Applications: Fix Cert C issues

+++ fsci/source/fsci_ble_gatt.c
@@ -6008,7 +6008,7 @@ static void HandleGattProcDiscoverAllPrimaryServices
     {
         /* Consider also the size of the kept services */
         dataSize += sizeof(uint8_t);
-        for(uint32_t iCount = 0; iCount < (uint8_t)*pArraySize; iCount++)
+        for(uint16_t iCount = 0U; iCount < *pArraySize; iCount++)
         {
             dataSize += fsciBleGattClientGetServiceBufferSize(&pServices[iCount]);
         }
@@ -6040,7 +6040,7 @@ static void HandleGattProcDiscoverAllPrimaryServices
         /* Set services array size parameter in the buffer */
         fsciBleGetBufferFromUint8Value((uint8_t)*pArraySize, pBuffer);
         /* Set services array parameter in the buffer */
-        for(uint32_t iCount = 0; iCount < (uint8_t)*pArraySize; iCount++)
+        for(uint16_t iCount = 0U; iCount < *pArraySize; iCount++)
         {
             fsciBleGattClientGetBufferFromService(&pServices[iCount], &pBuffer);
         }
@@ -6108,7 +6108,7 @@ static void HandleGattProcDiscoverPrimaryServicesByUuid
     {
         /* Consider also the size of the kept services */
         dataSize += sizeof(uint8_t);
-        for(uint32_t iCount = 0; iCount < (uint8_t)*pArraySize; iCount++)
+        for(uint16_t iCount = 0U; iCount < *pArraySize; iCount++)
         {
             dataSize += fsciBleGattClientGetServiceBufferSize(&pServices[iCount]);
         }
@@ -6140,7 +6140,7 @@ static void HandleGattProcDiscoverPrimaryServicesByUuid
         /* Set services array size parameter in the buffer */
         fsciBleGetBufferFromUint8Value((uint8_t)*pArraySize, pBuffer);
         /* Set services array parameter in the buffer */
-        for(uint32_t iCount = 0; iCount < (uint8_t)*pArraySize; iCount++)
+        for(uint16_t iCount = 0U; iCount < *pArraySize; iCount++)
         {
             fsciBleGattClientGetBufferFromService(&pServices[iCount], &pBuffer);
         }
@@ -6388,7 +6388,7 @@ static void HandleGattProcDiscoverCharacteristicByUuid
     {
         /* Consider also the size of the kept characteristics */
         dataSize += sizeof(uint8_t);
-        for(uint32_t iCount = 0; iCount < (uint8_t)*pArraySize; iCount++)
+        for(uint16_t iCount = 0U; iCount < *pArraySize; iCount++)
         {
             dataSize += fsciBleGattClientGetCharacteristicBufferSize(&pCharacteristics[iCount]);
         }
@@ -6420,7 +6420,7 @@ static void HandleGattProcDiscoverCharacteristicByUuid
         /* Set characteristics array size parameter in the buffer */
         fsciBleGetBufferFromUint8Value((uint8_t)*pArraySize, pBuffer);
         /* Set characteristics array parameter in the buffer */
-        for(uint32_t iCount = 0; iCount < (uint8_t)*pArraySize; iCount++)
+        for(uint16_t iCount = 0U; iCount < *pArraySize; iCount++)
         {
             fsciBleGattClientGetBufferFromCharacteristic(&pCharacteristics[iCount], &pBuffer);
         }
